refactor: renamed the controllers name to english
