+++ src/main/java/com/produtos/apirest/controller/UserController.java
@@ -16,11 +16,11 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/api/user")
-public class UsuarioController {
+public class UserController {
 
     private final UserService userService;
 
-    public UsuarioController(UserService userService) {
+    public UserController(UserService userService) {
         this.userService = userService;
     }
