+++ src/test/java/com/produtos/apirest/Controller/ExpertiseControllerTeste.java
@@ -1,9 +1,11 @@
 package com.produtos.apirest.Controller;
 
 import com.produtos.apirest.Util.Util;
-import com.produtos.apirest.models.DTO.OwnerDTO;
-import com.produtos.apirest.models.Owner;
-import com.produtos.apirest.service.OwnerService;
+import com.produtos.apirest.models.Area;
+import com.produtos.apirest.models.DTO.ExpertiseDTO;
+import com.produtos.apirest.models.Expertise;
+import com.produtos.apirest.service.AreaService;
+import com.produtos.apirest.service.ExpertiseService;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,53 +22,55 @@ import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 import java.util.ArrayList;
 import java.util.List;
 
-import static com.produtos.apirest.Controller.AnimalControllerTeste.generateAnimalList;
+import static com.produtos.apirest.Controller.AreaControllerTeste.generateAreaInstance;
 import static com.produtos.apirest.Util.Util.buildRequest;
 import static com.produtos.apirest.Util.Util.toJson;
-import static org.mockito.ArgumentMatchers.isA;
 
 @ActiveProfiles("test")
 @AutoConfigureMockMvc
 @SpringBootTest
-public class DonoControllerTeste {
+public class ExpertiseControllerTeste {
 
-    private final String API = "/api/owner";
+    private final String API = "/api/expertise";
 
     @MockBean
-    public OwnerService ownerService;
+    private ExpertiseService expertiseService;
+
+    @MockBean
+    private AreaService areaService;
 
     @Autowired
     MockMvc mvc;
 
-    protected static Owner generateOwner(){
-        return Owner.builder()
-                .ownerId(1L)
-                .name("name")
-                .cpf("cpf")
-                .phone("phone")
-                .build();
-    }
-
-    public static OwnerDTO generateDonoDTOInstance(){
-        return OwnerDTO.builder()
+    public static ExpertiseDTO generateEspecialidadeDTOInstance(){
+        return ExpertiseDTO.builder()
                 .id(1L)
+                .areaId(1L)
                 .name("name")
-                .cpf("cpf")
-                .phone("phone")
+                .area(generateAreaInstance())
                 .build();
     }
 
-    public static List<Owner> generateListDonoInstance(){
+    public static Expertise generateEspecialidadeInstance(){
+        return Expertise.builder()
+                .expertiseId(1L)
+                .name("name")
+                .area(generateAreaInstance())
+                .build();
+    }
+
+    public static List<Expertise> generateEspecialidadeListInstance(){
         return new ArrayList<>(){{
-            add(generateOwner());
+            add(generateEspecialidadeInstance());
         }};
     }
 
     @WithUserDetails("Admin")
     @Test
     public void deveSalvar() throws Exception{
-        Mockito.when(ownerService.save(Mockito.any(Owner.class))).thenReturn(generateOwner());
-        String json = toJson(generateDonoDTOInstance());
+        Mockito.when(expertiseService.save(Mockito.any(Expertise.class))).thenReturn(generateEspecialidadeInstance());
+        Mockito.when(areaService.save(Mockito.any(Area.class))).thenReturn(generateAreaInstance());
+        String json = toJson(generateEspecialidadeDTOInstance());
         MockHttpServletRequestBuilder request = Util.buildRequest(HttpMethod.POST, API, json);
         mvc.perform(request)
                 .andExpect(MockMvcResultMatchers.status().isCreated());
@@ -75,8 +79,8 @@ public class DonoControllerTeste {
     @WithUserDetails("Admin")
     @Test
     public void deveAtualizar() throws Exception{
-        Mockito.when(ownerService.update(Mockito.any(Owner.class))).thenReturn(generateOwner());
-        String json = toJson(generateDonoDTOInstance());
+        Mockito.when(expertiseService.update(Mockito.any(Expertise.class))).thenReturn(generateEspecialidadeInstance());
+        String json = toJson(generateEspecialidadeDTOInstance());
         MockHttpServletRequestBuilder request = Util.buildRequest(HttpMethod.PUT, API, json);
         mvc.perform(request)
                 .andExpect(MockMvcResultMatchers.status().isOk());
@@ -84,20 +88,20 @@ public class DonoControllerTeste {
 
     @WithUserDetails("Admin")
     @Test
-    public void deveRemoverPorId() throws Exception{
+    public void deveRemover() throws Exception{
         Long id = 1L;
-        Mockito.doNothing().when(ownerService).removeById(isA(Long.class));
+        Mockito.when(expertiseService.findById(Mockito.anyLong())).thenReturn(generateEspecialidadeInstance());
+        Mockito.doNothing().when(expertiseService).remove(Mockito.any(Expertise.class));
         MockHttpServletRequestBuilder request = buildRequest(HttpMethod.DELETE, API.concat("/").concat(String.valueOf(id)));
         mvc.perform(request)
-                .andExpect(MockMvcResultMatchers.status().isOk())
-                .andExpect(MockMvcResultMatchers.content().string("\"NO_CONTENT\""));
+                .andExpect(MockMvcResultMatchers.status().isOk());
     }
 
     @WithUserDetails("Admin")
     @Test
     public void deveRemoverComFeedback() throws Exception{
         Long id = 1L;
-        Mockito.when(ownerService.removeByIdWithFeedback(Mockito.anyLong())).thenReturn(generateOwner());
+        Mockito.when(expertiseService.removeByIdWithFeedback(Mockito.anyLong())).thenReturn(generateEspecialidadeInstance());
         MockHttpServletRequestBuilder request = buildRequest(HttpMethod.DELETE, API.concat("/feedback/").concat(String.valueOf(id)));
         mvc.perform(request)
                 .andExpect(MockMvcResultMatchers.status().isOk());
@@ -105,9 +109,9 @@ public class DonoControllerTeste {
 
     @WithUserDetails("Admin")
     @Test
-    public void deveBuscarComFiltro() throws Exception{
-        Mockito.when(ownerService.find(Mockito.any(Owner.class))).thenReturn(generateListDonoInstance());
-        String json = toJson(generateDonoDTOInstance());
+    public void deveBuscarPorFiltro() throws Exception{
+        Mockito.when(expertiseService.find(Mockito.any(Expertise.class))).thenReturn(generateEspecialidadeListInstance());
+        String json = toJson(generateEspecialidadeDTOInstance());
         MockHttpServletRequestBuilder request = Util.buildRequest(HttpMethod.GET, API.concat("/filter"), json);
         mvc.perform(request)
                 .andExpect(MockMvcResultMatchers.status().isOk());
@@ -117,7 +121,7 @@ public class DonoControllerTeste {
     @Test
     public void deveBuscarPorId() throws Exception{
         Long id = 1L;
-        Mockito.when(ownerService.findById(Mockito.anyLong())).thenReturn(generateOwner());
+        Mockito.when(expertiseService.findById(Mockito.anyLong())).thenReturn(generateEspecialidadeInstance());
         MockHttpServletRequestBuilder request = buildRequest(HttpMethod.GET, API.concat("/").concat(String.valueOf(id)));
         mvc.perform(request)
                 .andExpect(MockMvcResultMatchers.status().isOk());
@@ -126,18 +130,8 @@ public class DonoControllerTeste {
     @WithUserDetails("Admin")
     @Test
     public void deveBuscarTodos() throws Exception{
-        Mockito.when(ownerService.findAll()).thenReturn(generateListDonoInstance());
+        Mockito.when(expertiseService.findAll()).thenReturn(generateEspecialidadeListInstance());
         MockHttpServletRequestBuilder request = buildRequest(HttpMethod.GET, API);
-        mvc.perform(request)
-                .andExpect(MockMvcResultMatchers.status().isOk());
-    }
-
-    @WithUserDetails("Admin")
-    @Test
-    public void deveBuscarTodosAnimais() throws Exception{
-        Long id = 1L;
-        Mockito.when(ownerService.findAllAnimalsByOwnerId(Mockito.anyLong())).thenReturn(generateAnimalList());
-        MockHttpServletRequestBuilder request = buildRequest(HttpMethod.GET, API.concat("/animals/".concat(String.valueOf(id))));
         mvc.perform(request)
                 .andExpect(MockMvcResultMatchers.status().isOk());
     }
